Improve error and event history rendering (#80)
* perf(errors): progressively render error cards

Render four cards synchronously and append remaining active/history cards in LVGL timer batches. Tighten heavy error-screen initial render budget to 150 ms.


* fix(errors): progressively render every error

Remove the legacy active/history display caps and summary glyphs so all errors remain accessible while loading in batches.


* feat(events): group entries by date

Insert localized Today, Yesterday, and calendar-date separators while preserving lazy loading across batch boundaries.


* fix(events): backfill timestamps after NTP sync

Tag events with a per-boot ID, use 64-bit uptime, migrate persisted v1 history, and backfill only untimed entries from the current boot when accurate time arrives.


---------

+++ main/api_server.cpp
@@ -3098,7 +3098,7 @@ static esp_err_t events_get_handler(httpd_req_t* req)
         cJSON* evt = cJSON_CreateObject();
         cJSON_AddStringToObject(evt, "type", event_type_name(events[i].type));
         cJSON_AddNumberToObject(evt, "timestamp", events[i].timestamp);
-        cJSON_AddNumberToObject(evt, "uptime_ms", events[i].uptime_ms);
+        cJSON_AddNumberToObject(evt, "uptime_ms", (double)events[i].uptime_ms);
         cJSON_AddNumberToObject(evt, "payload", events[i].payload);
         cJSON_AddItemToArray(arr, evt);
     }
